Removed warning
git-svn-id: https://svn.wxwidgets.org/svn/wx/wxWidgets/trunk@25854 c3d73ce0-8a6f-49c7-b76d-6d57e0e08775

+++ src/generic/dirctrlg.cpp
@@ -1680,7 +1680,7 @@ int wxFileIconsTable::GetIconID(const wxString& extension, const wxString& mime)
     const unsigned int size = 16;
     
     int id = m_smallImageList->GetImageCount();
-    if ((bmp.GetWidth() == size) && (bmp.GetHeight() == size))
+    if ((bmp.GetWidth() == (int) size) && (bmp.GetHeight() == (int) size))
     {
         m_smallImageList->Add(bmp);
     }
